# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.HTTPError:
    properties:
      error:
        type: string
    type: object
...
  models.IncomeStatement:
    properties:
      by_function_administrative_expenses:
        type: string
      by_function_cost_of_goods_sold:
        type: string
      by_function_gross_profit:
        type: string
      by_function_other_operating_revenues:
        type: string
      by_function_selling_expenses:
        type: string
      by_nature_depreciation_expenses:
        type: string
      by_nature_inventory_change:
        type: string
      by_nature_labour_expenses:
        type: string
      by_nature_long_term_investment_expenses:
        type: string
      by_nature_material_expenses:
        type: string
      by_nature_other_operating_revenues:
        type: string
      equity_investment_earnings:
        type: string
      extra_dividends:
        type: string
      extra_expenses:
        type: string
      extra_revenues:
        type: string
      file_id:
        type: string
      id:
        type: integer
      income_after_income_taxes:
        type: string
      income_before_income_taxes:
        type: string
      interest_expenses:
        type: string
      investment_fair_value_adjustments:
        type: string
      net_income:
        type: string
      net_turnover:
        type: string
      other_interest_revenues:
        type: string
      other_long_term_investment_earnings:
        type: string
      other_operating_expenses:
        type: string
      other_taxes:
        type: string
      provision_for_income_taxes:
        type: string
      statement_id:
        type: string
    type: object
  models.Member:
    properties:
      at_legal_entity_registration_number:
        type: string
      birth_date:
        type: string
      date_from:
        type: string
      entity_type:
        type: string
      id:
        type: integer
      last_modified_at:
        type: string
      latvian_identity_number_masked:
        type: string
      legal_entity_registration_number:
        type: string
      name:
        type: string
      number_of_shares:
        type: string
      registered_on:
        type: string
      share_currency:
        type: string
      share_nominal_value:
        type: string
      uri:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for managing data based on the provided schema.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Your API Title
  version: "1.0"
paths:
...
  /income-statements:
    get:
      description: Retrieve a list of all income statement records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.IncomeStatement'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get all income statement entries
      tags:
      - income-statements
    post:
      consumes:
      - application/json
      description: Add a new income statement record to the database
      parameters:
      - description: Income Statement data to create
        in: body
        name: incomeStatement
        required: true
        schema:
          $ref: '#/definitions/models.IncomeStatement'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IncomeStatement'
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Create a new income statement entry
      tags:
      - income-statements
  /income-statements/{id}:
    delete:
      description: Remove an income statement record from the database using its ID
      parameters:
      - description: Income Statement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Income Statement not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Delete an income statement entry by ID
      tags:
      - income-statements
    get:
      description: Retrieve details of a specific income statement record using its
        ID
      parameters:
      - description: Income Statement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IncomeStatement'
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Income Statement not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get a single income statement entry by ID
      tags:
      - income-statements
    put:
      consumes:
      - application/json
      description: Modify the details of an existing income statement record by ID
      parameters:
      - description: Income Statement ID
        in: path
        name: id
        required: true
        type: integer
      - description: Income Statement data to update
        in: body
        name: incomeStatement
        required: true
        schema:
          $ref: '#/definitions/models.IncomeStatement'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IncomeStatement'
        "400":
          description: Bad Request - Invalid ID format or input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Income Statement not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Update an existing income statement entry
      tags:
      - income-statements
  /members:
    get:
      description: Retrieve a list of all member records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Member'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get all member entries
      tags:
      - members
    post:
      consumes:
      - application/json
      description: Add a new member record to the database
      parameters:
      - description: Member data to create
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/models.Member'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Member'
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Create a new member entry
      tags:
      - members
  /members/{id}:
    delete:
      description: Remove a member record from the database using its ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Member not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Delete a member entry by ID
      tags:
      - members
    get:
      description: Retrieve details of a specific member record using its ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Member not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get a single member entry by ID
      tags:
      - members
    put:
      consumes:
      - application/json
      description: Modify the details of an existing member record by ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      - description: Member data to update
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/models.Member'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
        "400":
          description: Bad Request - Invalid ID format or input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Member not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Update an existing member entry
      tags:
      - members
...
schemes:
- http
- https
swagger: "2.0"